feat: Controller View 클래스 annotation 방식으로 변경
